refactor NIP-46 impl.
