+++ src/index.ts
@@ -1,5 +1,5 @@
 export type { NostrSigner } from "./interface";
 export { Nip07ExtensionSigner, type Nip07Extension } from "./nip07";
-export { Nip46RemoteSigner } from "./nip46";
-export type { Nip46ClientMetadata, Nip46ConnectionParams, Nip46SessionState } from "./nip46";
+export { Nip46RemoteSigner } from "./nip46/signer";
+export type { Nip46ClientMetadata, Nip46ConnectionParams, Nip46SessionState } from "./nip46/signer";
 export { SecretKeySigner } from "./secret_key";
